+++ CHANGELOG.md
@@ -1,15 +1,29 @@
-# Party Marker Classic
+# Party Marker
+
+## v1.2.4
+
+- Just cleanup for automatic pushing
+
+## v1.2.3
+
+- Updated for new API calls
+- Support for 1 or 2 partners in the group
+- Including classic and retail versions
 
 ## v1.2.2
+
 - Pulled loot controls out of the marking control block (if marking fails for some reason, loot can still be set)
 - Fixed loot controls not working
 - Cleaned up the addon's messaging system
 
 ## v1.2.1
+
 - Adjust .toc file and curseforge management stuff
 
 ## v1.2.0
+
 - Options now found in interface/addons panel using ace3
 
 ## v1.1.1
+
 - Fixed issues with war mode